Rename sdmmc as sdcard.
Matches the name of the primary type contained within it.
